% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_names.R
\name{add_r.n}
\alias{add_r.n}
\title{Takes result of kern_2_df and changes
rhythms to name values (ie changes 4 to quarter note)}
\usage{
add_r.n(ri)
}
\arguments{
\item{ri}{individual note vector}
}
\value{
columns with rhythm names
}
\description{
Takes result of kern_2_df and changes
rhythms to name values (ie changes 4 to quarter note)
}
