% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_to_df.R
\name{all_info_piece}
\alias{all_info_piece}
\title{Takes a piece_df from piece_df()}
\usage{
all_info_piece(piece_df)
}
\value{
list of df of notes, key, time base etc.
}
\description{
Takes a piece_df from piece_df()
}
