% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music-utils.R
\name{calc_12_tone_abs_freq}
\alias{calc_12_tone_abs_freq}
\title{Calculate the absoulte frequency based on the distance
to/from a given reference pitch
p reference pitch (usually 440)
d distance from reference pitch (in semitones)}
\usage{
calc_12_tone_abs_freq(d, p = 440)
}
\description{
Calculate the absoulte frequency based on the distance
to/from a given reference pitch
p reference pitch (usually 440)
d distance from reference pitch (in semitones)
}
