% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music-utils.R
\name{calc_n_tone_abs_freq}
\alias{calc_n_tone_abs_freq}
\title{Calculate the absoulte frequency based on the distance
to/from a given reference pitch}
\usage{
calc_n_tone_abs_freq(n, d, p = 440)
}
\arguments{
\item{n}{base integer for n-tone system}

\item{d}{distance from reference pitch (in semitones)}

\item{p}{reference pitch (usually 440)}
}
\description{
Calculate the absoulte frequency based on the distance
to/from a given reference pitch
}
