% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music-utils.R
\name{compute_nth_tone_ratio}
\alias{compute_nth_tone_ratio}
\title{d distance from root (origin)
n nth root (base of system e.g. 12)}
\usage{
compute_nth_tone_ratio(d, n)
}
\description{
d distance from root (origin)
n nth root (base of system e.g. 12)
}
