% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melodic_intervals.R
\name{consonances}
\alias{consonances}
\title{consonances}
\usage{
consonances(piece, col)
}
\arguments{
\item{piece}{One piece}

\item{col}{name of intstrument}
}
\value{
vector of counts for consonances
}
\description{
consonances
}
