% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{edges}
\alias{edges}
\title{Euler's relationship to Schlafli numbers for convex polytopes to find E}
\usage{
edges(p, q)
}
\description{
Euler's relationship to Schlafli numbers for convex polytopes to find E
}
