% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\name{enum}
\alias{enum}
\title{create an enum S3 class given vector of strings and a enum classname}
\usage{
enum(enums, name = NULL)
}
\arguments{
\item{enums}{list of (symbol, value) mappings for enumerable class}

\item{name}{(optional) name of enumerable class}
}
\description{
Factory function enum that takes a vector of strings and a name and  returns
  an enum object. Dispatches on case-insensitive indexing argument list
}
\examples{

# col <- enum(c("BLUE", "red", "Black"), name = "TheColors")
col$R
#> [1] "RED"
class(col)
#> [1] "TheColors" "enum"      "list"

side <- enum(c("left", "right"))
side$L
#> [1] "LEFT"
class(side)
#> [1] "enum" "list"
}
