% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music-utils.R
\name{equal_temperament_intervals}
\alias{equal_temperament_intervals}
\title{Returns direct ratios of 12-tone equal temperament}
\usage{
equal_temperament_intervals()
}
\description{
Returns direct ratios of 12-tone equal temperament
}
