% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{faces}
\alias{faces}
\title{Uses Euler's relationship to Schlafli numbers for convex polytopes to find F}
\usage{
faces(p, q)
}
\description{
Uses Euler's relationship to Schlafli numbers for convex polytopes to find F
}
