% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier.R
\name{fourier_magnitude}
\alias{fourier_magnitude}
\title{Naive fourier (magnitude plot) transform function to plot magnitude}
\usage{
fourier_magnitude(ft = example_ft, w_size = 50, precision = 0.01)
}
\description{
Naive fourier (magnitude plot) transform function to plot magnitude
}
