% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music-utils.R
\name{get_solfeggio_scale}
\alias{get_solfeggio_scale}
\title{Apply ratios to a given base frequency (Default: 174Hz)}
\usage{
get_solfeggio_scale(extended = TRUE, base = if (extended) 174 else 396)
}
\description{
Apply ratios to a given base frequency (Default: 174Hz)
}
