% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{in_range}
\alias{in_range}
\title{returns whether x is within a specific delta range of y.}
\usage{
in_range(x, y, err = 1)
}
\description{
returns whether x is within a specific delta range of y.
}
\examples{
 in_range(4, 6, err = 2)
}
