% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{is_list}
\alias{is_list}
\title{Is x explicitly a list type}
\usage{
is_list(x)
}
\arguments{
\item{x}{input objects}
}
\description{
Is x explicitly a list type
}
\examples{
is_list(c(1,2,3))
is_list(list(99))
}
