% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melodic_intervals.R
\name{is_mel_int}
\alias{is_mel_int}
\title{is_mel_int
Checks if two notes is a specified melodic interval}
\usage{
is_mel_int(n1, n2, int)
}
\arguments{
\item{n1}{one note}

\item{n2}{second note}

\item{int}{what interval we are chekcing for}
}
\value{
T or F
}
\description{
is_mel_int
Checks if two notes is a specified melodic interval
}
