% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{kern2data}
\alias{kern2data}
\title{Takes a .krn file and converts into a readable format}
\usage{
kern2data(spline)
}
\arguments{
\item{spline}{.krn file}
}
\value{
dataframe of stuff
}
\description{
Takes a .krn file and converts into a readable format
}
