% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_to_df.R
\name{kern2df}
\alias{kern2df}
\title{Takes a .krn spline and converts into data frame}
\usage{
kern2df(spline)
}
\arguments{
\item{spline}{.krn file}
}
\value{
dataframe of stuff
}
\description{
Takes a .krn spline and converts into data frame
}
