% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melodic_intervals.R
\name{mel_ints}
\alias{mel_ints}
\title{voice_mel_ints}
\usage{
mel_ints(piece, col)
}
\arguments{
\item{piece}{One instrument}

\item{col}{name}
}
\value{
vector of counts for melodic intervals
}
\description{
voice_mel_ints
}
