% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_names.R
\name{n.v_n.n}
\alias{n.v_n.n}
\title{Identifies the note and note value of a .krn note}
\usage{
n.v_n.n(note)
}
\arguments{
\item{note}{.krn file}
}
\value{
NNV and DNV
}
\description{
Identifies the note and note value of a .krn note
}
