% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{n_primes}
\alias{n_primes}
\title{Compute n primes}
\usage{
n_primes(length.out)
}
\description{
Try to keep it under 5000, or your CPU might explode
}
\details{
Microbenchmark for n_primes (microseconds):
     expr             min         lq       mean
n_primes(10)       72.402       75.441   82.7702
n_primes(100)     2687.685     2852.065  2945.43
n_primes(1000)   417199.246   417258.025 426916
n_primes(2000)  1860778.072  1924292.325 1992891
n_primes(3000)  4366437.459  4403578.398 4413822
n_primes(5000) 13054581.632 13071920.879 13494860
}
