% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music-utils.R
\name{nearest_pitch}
\alias{nearest_pitch}
\title{Find closest pitch class symbol for an incoming frequency}
\usage{
nearest_pitch(freq, base = 440)
}
\description{
Find closest pitch class symbol for an incoming frequency
}
