% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{note_duration}
\alias{note_duration}
\title{note_duration}
\usage{
note_duration(piece, inst)
}
\arguments{
\item{piece}{in raw data frame}
}
\value{
density for each measure
}
\description{
note_duration
}
