% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{nroot}
\alias{nroot}
\title{Compute bth root of a
This is a so-called Vectorized function, i.e.
f(A,B) == c(f(A[1],B[1]), f(A[2],B[2]), ...) for vectors A, B
of equal length.}
\usage{
nroot(a, b)
}
\description{
This does *not* handle the case where b is the inverse of an integer
(e.g. nthroot(-1,1/3) ==? (-1)^3), since in general you cannot count
on 1/(1/int) being an integer.
If on the other hand you want the complex principal value of the nth
root, you can use (a+0i)^(1/b).
}
