% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_to_df.R
\name{piece_df}
\alias{piece_df}
\title{piece_df 
Takes splines in .krn and calls kern2df 
to create one df for entire piece}
\usage{
piece_df(v, insts)
}
\arguments{
\item{v}{vector with splines .krn file strings}

\item{insts}{vector of same lenght as v, 
including instrument names for each spline}
}
\value{
dataframe for entire piece - combined splines
}
\description{
piece_df 
Takes splines in .krn and calls kern2df 
to create one df for entire piece
}
