% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_to_df.R
\name{piece_df2}
\alias{piece_df2}
\title{Takes splines in .krn and calls kern2df to create one df for entire piece}
\usage{
piece_df2(v, insts)
}
\arguments{
\item{v}{vector with splines .krn file strings}

\item{insts}{vector of same lenght as v, including instrument names
for each spline}
}
\value{
dataframe for entire piece - combined splines
}
\description{
Takes splines in .krn and calls kern2df to create one df for entire piece
}
