% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchclass.R
\name{Pitch.Hertz}
\alias{Pitch.Hertz}
\title{initialize a pitch class enum with a base frequency (e.g. A == 440)}
\usage{
Pitch.Hertz(base = 440)
}
\description{
initialize a pitch class enum with a base frequency (e.g. A == 440)
}
