% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tone-circle.R
\name{pitch_to_angle}
\alias{pitch_to_angle}
\title{Return angle value for pitch class on tone circle}
\usage{
pitch_to_angle(pitch, orientation = "chromatic")
}
\description{
Return angle value for pitch class on tone circle
}
