% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchclass.R
\name{PitchAngle.C}
\alias{PitchAngle.C}
\title{Enum containing pitch angles, chromatic tone circle (counter-clockwise)}
\usage{
PitchAngle.C()
}
\description{
Enum containing pitch angles, chromatic tone circle (counter-clockwise)
}
\details{
returned angles are in 30 degree increments along the polar axis
}
