% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_degree.R
\name{scale_degree_freq}
\alias{scale_degree_freq}
\title{scale_degree_freq}
\usage{
scale_degree_freq(piece)
}
\arguments{
\item{piece}{}
}
\value{
the frequency of occurance of all the scale degrees
}
\description{
scale_degree_freq
}
