% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music-utils.R
\name{solfeggio_ratios}
\alias{solfeggio_ratios}
\title{Get numeric ratio for solfeggio scale}
\usage{
solfeggio_ratios(extended = TRUE, return_both = FALSE)
}
\description{
Get numeric ratio for solfeggio scale
}
