% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tone-circle.R
\name{tone_circle}
\alias{tone_circle}
\title{Create a new tone circle.}
\usage{
tone_circle(v, base = 440, n_tones = 12)
}
\arguments{
\item{v}{Vertices of the tone circle}

\item{base}{Base frequency for initial pitch (e.g. 440)}

\item{n_tones}{n-tone system.  (default 12 for 12-tone system)}
}
\description{
Create a new tone circle.
}
