% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music-utils.R
\name{tone_random_walk}
\alias{tone_random_walk}
\title{Generate an n-tone random walk sequence of intervals (equal temperament)}
\usage{
tone_random_walk(n = 12, start_pitch = 440, length.out = 100, weights = NULL)
}
\description{
Generate an n-tone random walk sequence of intervals (equal temperament)
}
