% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music-utils.R
\name{tone_reference_ratios}
\alias{tone_reference_ratios}
\title{Return n-tone ratios based on a reference pitch}
\usage{
tone_reference_ratios(n, base)
}
\description{
Return n-tone ratios based on a reference pitch
}
