% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melodic_intervals.R
\name{top_line}
\alias{top_line}
\title{top_line}
\usage{
top_line(piece, inst)
}
\arguments{
\item{piece}{}

\item{col}{one instrument name name}
}
\value{
gives the top melodic line of a instrument
}
\description{
top_line
}
