% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melodic_intervals.R
\name{top_line2}
\alias{top_line2}
\title{top_line2}
\usage{
top_line2(piece, inst)
}
\arguments{
\item{piece}{piece of krn}

\item{col}{one instrument name name}
}
\value{
gives the top melodic line of a instrument
}
\description{
top_line2
}
