% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melodic_intervals.R
\name{tot_mel_int}
\alias{tot_mel_int}
\title{tot_mel_int
How many melodic intervals each spline has}
\usage{
tot_mel_int(spline)
}
\arguments{
\item{spline}{.krn file - grouped by spline number}
}
\value{
how many melodic intervals each spline has
}
\description{
tot_mel_int
How many melodic intervals each spline has
}
\examples{


}
