% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music-utils.R
\name{twelve_tone_refernce_ratios}
\alias{twelve_tone_refernce_ratios}
\title{Return 12-tone ratios based on a reference pitch}
\usage{
twelve_tone_refernce_ratios(base)
}
\description{
Return 12-tone ratios based on a reference pitch
}
