% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier.R
\name{untwist.one}
\alias{untwist.one}
\title{Untwist one (fourier component) for a specific frequency}
\usage{
untwist.one(freq, .rotations)
}
\description{
Untwist one (fourier component) for a specific frequency
}
