% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier.R
\name{untwist.range}
\alias{untwist.range}
\title{Use untwist.one to iterate through a range of rotation values for a "moving
image" feel.}
\usage{
untwist.range(freq, rotation_range = seq(10, 250, 7))
}
\description{
Use untwist.one to iterate through a range of rotation values for a "moving
image" feel.
}
