% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier.R
\name{untwist.viz}
\alias{untwist.viz}
\title{TODO: Add gganimate or googleVis to concat frames into a video
TODO: add circle diagrams for plotting (unit circle for tone diagrams)
Plot untwisting motion of a given frequency in the complex plane
USE Plotly for polar charts to mimic tone circles}
\usage{
untwist.viz(X, Y, .F, N = length(X))
}
\description{
TODO: Add gganimate or googleVis to concat frames into a video
TODO: add circle diagrams for plotting (unit circle for tone diagrams)
Plot untwisting motion of a given frequency in the complex plane
USE Plotly for polar charts to mimic tone circles
}
