% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{vertices}
\alias{vertices}
\title{Euler's relationship to Schlafli numbers for convex polytopes to find V}
\usage{
vertices(p, q)
}
\description{
Euler's relationship to Schlafli numbers for convex polytopes to find V
}
