% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_names.R
\name{add_n.v_n.n}
\alias{add_n.v_n.n}
\title{add_n.v_n.n
Add note value note name
Adds columns with more easily analyzable note names}
\usage{
add_n.v_n.n(notes)
}
\arguments{
\item{notes}{one note line for one instrument}
}
\value{
data frame with orriginal piece and NNV and DNV
}
\description{
add_n.v_n.n
Add note value note name
Adds columns with more easily analyzable note names
}
